### 2019-10-04 #1781 @kyleniemeyer
Hi @cimentadaj, thanks for your submission. Before we get started, can you please rename the article source file to `paper.md`? Our automated build system requires the article to be called that, though it can be located anywhere in the repo.
