### 2019-10-07 7779c16
Renames paper to paper.*
