### 2019-10-08 #1781 @briatte
Dear all,

I'd be happy to review the package, although let me admit that I'm biased, since I know @cimentadaj's excellent work on other survey-related packages :)

All the best,

François
