### 2019-10-10 #1781 @majensen
@amoeba, would you be willing to review [this submission](https://github.com/openjournals/joss-papers/blob/joss.01781/joss.01781/10.21105.joss.01781.pdf) to JOSS? Thanks for considering!
