### 2019-10-10 #1796 @whedon
[REVIEW]: perccalc: An R package for estimating percentiles from categorical variables
**Submitting author:** @cimentadaj (<a href="http://orcid.org/0000-0001-5594-1156">Jorge Cimentada</a>)
**Repository:** <a href="https://github.com/cimentadaj/perccalc/" target ="_blank">https://github.com/cimentadaj/perccalc/</a>
**Version:** v1.0.5
