### 2019-10-19 cimentadaj/perccalc#2 @amoeba
JOSS Review
Hey @cimentadaj, I've had a chance to sit down with `perccalc`
 and have checked off what I could on the checklist over on https://github.com/openjournals/joss-reviews/issues/1796.

Overall, I think you've done a really nice job with this package, both code, documentation, and tests. Notably, I was immensely appreciative as a reviewer to see you copy in Reardon's code from his implementation for verification, especially as I wasn't able to get my hands on original appendix.

I left a few things unchecked and you'll see why below. Please take a look and let me know if I can explain anything.

### General Checks

- License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

	No, though what's been done is reasonable for an R package. Please change the text of the `LICENSE` file to contain the text of an [OSI Approved license](https://opensource.org/licenses/alphabetical) per the [Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).

### Documentation

(JOSS states that the README at the root should check these boxes. You check these boxes in other places so these are mostly about moving some info to the README)

- A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

  Sort of. I think the initial paragraph could be stronger if it had a sentence explaining _why_ someone would want to use this package. The sentence would be intended for a person who doesn't yet know about the method. The paper does a nice job tackling this so maybe something from that could be borrowed.

- Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

  An example is given but with synthetic data. I think it'd be great to try to find a real-world dataset to run the two functions on if possible. Would the GSSvocab dataset work here? The package includes examples with real data in other places and any one of these would be great candidates and would make the README stronger.

### Paper

Overall the paper reads great and I appreciate the time you put into it, it's excellent.

- The paper includes detail in the Examples section that JOSS guidelines indicate should be removed, see [Guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html):

  > "Note the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.". Please remove this section.
- The footer has a malformed citation
- The references section needs some DOIs:
  - The reference to Bassok et al. needs a DOI of https://doi.org/10.1177/2332858416653924
  - The reference of Chmielewski and Reardon needs a DOI of https://doi.org/10.1177/2332858416649593
  - The reference of Erikson et al. needs a DOI of https://doi.org/10.2307/589632
  - The reference of Reardon and Portilla needs its DOI of https://doi.org/10.1177/2332858416657343

### Other

I had some other suggestions which are not specific to parts of the JOSS review and can be considered optional:

- I got a warning on this section of the example in the README:

  > perc_diff(df, categorical, continuous, percentiles = c(90, 10))
  difference          se
  79.61141798  0.08179896
  Warning message:
  `cols` is now required.
  Please use `cols = c(value)`

  Could this warning be avoided?

- This errors out in an odd way and could be a more descriptive errors if you think it'd help users:

> perc_diff(df, foo, bar, percentiles = c(90, 10))
Error in category_summary(data_model, variable_name, continuous_name,  :
  is_ordered_fct is not TRUE

  - Typo in "A vector with the percentile difference and it's associated standard error" in perc_diff. "it's"  should be "its".

- In the README, "Reardon (2011)" could be hyperlinked. It's down at the bottom of the readme but I like links when I'm on websites.
- In Usage section of README, the first code block doesn't list `library(perccalc)`. I get that this is listed above in the Installation section but I think it might be more clear to repeat it in this code block
