### 2019-10-19 #1796 @amoeba
Hey @majensen, @cimentadaj, I've conducted my first pass at the review checklist and things look very good overall. I have some items I'd like @cimentadaj to address so at this point my review is a conditional accept. See: https://github.com/cimentadaj/perccalc/issues/2.
