### 2019-10-20 2154fde
Update paper.Rmd
Typos.
