### 2019-10-20 cimentadaj/perccalc#2 @briatte
Dear all,

Here's my own review.

As I stated when I accepted the review, I know @cimentadaj's high-quality work on other packages, and this one does not disappoint either: it's a very good package that is very clear about what it does and how to use it.

The docs are also excellent, esp. considering that they also document [borderline cases](https://cimentadaj.github.io/perccalc/articles/perc_warning_example.html) where the package is unable to produce all its results.

~~I'm still checking various aspects of the package, but here are some points already.~~ I'm done with the review, for now. All points below are minor, so conditional accept, with very minor revisions.

## General checks

### License
I have left this item unchecked after reading @amoeba's [comment](#issue-509374037), which points to the fact that the [current LICENSE](https://github.com/cimentadaj/perccalc/blob/master/LICENSE) is nonstandard.

The DESCRIPTION file indicates "MIT + file LICENSE", [as recommended by Wickham](http://r-pkgs.had.co.nz/description.html#license), but I guess the LICENSE file needs to include [the actual license](https://opensource.org/licenses/MIT).

I've submitted a PR: https://github.com/cimentadaj/perccalc/pull/3

## Functionality

### Installation
… went perfectly fine on my end, after a few dependency updates.

Suggestion -- would it be better to use `remotes` instead of `devtools` (which uses `remotes` internally) in the installation code?

Also -- like @amoeba, I favour having installation code in the "Install" section of the README, and then have the "Usage" section start with a `library` call.

## Documentation

### Community guidelines
Question -- is it now standard to have the CONTRIBUTING guide in `.github` instead of in the root of the repo?

## Software paper

I've submitted some PRs to it:
- https://github.com/cimentadaj/perccalc/pull/4 (suggested change in first sentence)
- https://github.com/cimentadaj/perccalc/pull/5 (typos)
- https://github.com/cimentadaj/perccalc/pull/6 (code suggestion)

More borderline-pathological nitpicking below.

1.
The code in the paper won't run on old versions of `tidyr` -- specifically, line 145 (copied below) will fail on e.g. `tidyr` 0.8.3, which is what I had installed before updating it to 1.0.0.

```r
nest(data = c(-year, -CNT))
```

2.
> (values of a continuous variable for the 0, 10th, 20th, ..., 100th percentile of the ordered categorical variable)

… That should be either 0, 10th, …, 99th *or* 1st, 10th, …, 100th -- right?

3.
> Mathematics

… Do you need the capital M in English?

### State of the field

I have not yet checked that box because, unless I missed a paragraph somewhere in the paper, the package does not refer to other packages dedicated to working with categorical variables, e.g. [`qualvar`](https://cran.r-project.org/package=qualvar) (disclosure: the author is a friend of mine).

Perhaps it might be worth mentioning a few such packages? Some of them show up in the [CRAN Task View: Official Statistics & Survey Methodology](https://cran.r-project.org/view=OfficialStatistics) -- where `perccalc` should also be listed IMO.

### References

I see that @amoeba has provided the almost-inevitable "missing DOI links" parts of the reviewing process. Thanks to him, since doing so is rather tedious.

To @majensen and JOSS editor(s) -- This happens on many reviews, right? I remember doing the same as @ameoba on a different review. Any way this could be further automated?

## Other

1.

While replicating the README, I was unable to replicate the warning message reported by @amoeba in [his review](#issue-509374037) ("Other" section).

2.

The package has a startup message:

```
Please cite as:

Cimentada, Jorge (2019). Estimate Percentiles from an Ordered Categorical Variable R package version 1.0.3.
```

I understand the point of having the message, but am not a big fan of startup messages overall.
