### 2019-10-21 #1796 @briatte
Hi @majensen, @cimentadaj and @amoeba

I'm also done with [my review](https://github.com/cimentadaj/perccalc/issues/2#issuecomment-544239961): conditional accept with very minor revisions (some of which are nitpicky to the point that @cimentadaj might reasonably decide to ignore them entirely).
