### 2019-10-22 #1796 @cimentadaj
Hi everyone

Thanks for the reviews. I'm probably gonna check them out first week of November and finish them at that time. Is that good?
