### 2019-11-07 cimentadaj/perccalc#2 @majensen
@briatte sorry just saw your comment re DOI links- I can push your suggestion up the chain; thanks!
