### 2019-11-08 9cb93a6
Adds correct MIT license
