### 2019-11-08 7b547a4
Adds better explanation of the package in README
