+++ README.Rmd
@@ -23,7 +23,10 @@ options(tibble.print_min = 5, tibble.print_max = 5)
 
 ## Overview
 
-Reardon (2011) introduced a very interesting concept in which he calculates percentile differences from ordered categorical variables. He explains his procedure very much in detail in the appendix of the book chapter but no formal implementation has been yet available on the web. With this package I introduce two functions that apply the procedure.
+`perccalc` is a direct implementation of the theoretical work of Reardon (2011) where it is possible to estimate the difference between two percentiles from an ordered categorical variable. More
+concretely, by specifying an ordered categorical variable and a continuous variable, this method can estimate differences in the continuous variable between percentiles of the ordered categorical
+variable. This brings forth a relevant strategy to contrast ordered categorical variables which
+usually have alternative continuous measures to the percentiles of the continuous measures. Moreover, this opens an avenue for calculating percentile distributions and percentile differences for ordered categorical variables which don't necessarily have an alternative continuous measure such as job occupation classifications. With this package I introduce two functions that apply the procedure.
 
 The package has two main functions:
 
@@ -35,9 +38,7 @@ The package has two main functions:
 You can install and load the package with these commands:
 
 ```{r, eval = FALSE}
-# install.packages("perccalc") # for stable version
-library(perccalc)
-
+install.packages("perccalc") # for stable version
 # or
 devtools::install_github("cimentadaj/perccalc") # for development version
 ```
@@ -45,11 +46,9 @@ devtools::install_github("cimentadaj/perccalc") # for development version
 ## Usage
 
 Suppose we have a dataset with one continuous variable and one categorical variable:
-```{r, echo = FALSE, message = FALSE}
-library(perccalc)
-```
 
 ```{r, message = FALSE, warning = FALSE}
+library(perccalc)
 library(dplyr)
 library(ggplot2)
 
