### 2019-11-08 0c1b11d
Makes README example more applied by using the GSS data
