+++ vignettes/perc_calculator_example.Rmd
@@ -41,7 +41,7 @@ This is the minimum dataset that the function will accept. This means that it ne
 
 The package is called `perccalc`, short for percentile calculator and we can install and load it with this code:
 
-```{r, eval = F}
+```{r, eval = FALSE}
 devtools::install_github("cimentadaj/perccalc")
 library(perccalc)
 ```
@@ -96,7 +96,7 @@ gss %>%
 
 Running the code below using the `gss_data.dta`..
 
-```{r, eval = F}
+```{r, eval = FALSE}
 *--------
 use "/Users/cimentadaj/Downloads/gss_data.dta", clear
 
