### 2019-11-08 9f8e829
Removes pkg API details from the paper
