+++ paper/paper.md
@@ -142,10 +142,8 @@ perc_diff(data_model = pisa,
 
 This means that the difference in Mathematics test scores between the 90th and
 10th percentile of father's education is 45 points with a standard error of 14
-points. We can extrapolate this example for each country separately using
-`dplyr::group_by`. Since the result of `perc_diff` is a vector, to be able to
-work seamlessly with `dplyr::group_by` we will use `perc_diff_df` which returns
-a data frame:
+points. We can extrapolate this example for each country separately using `perc_diff_df`
+which returns a data frame:
 
 
 ```r
@@ -182,9 +180,8 @@ It looks like Estonia has a much smaller achievement gap relative to Spain and
 Germany but also note that both Germany and Spain have been decreasing their
 inequality.
 
-`percalc` also has the `perc_dist` function which calculates the distribution of
-the percentiles: this is useful for comparing more finegrained percentiles
-rather than differences:
+In contrast, `perc_dist` calculates the distribution of the percentiles which
+is useful for comparing more finegrained percentiles rather than differences.
 
 
 ```r
