+++ paper/paper.bib
@@ -14,7 +14,8 @@
   number={3},
   pages={2332858416657343},
   year={2016},
-  publisher={SAGE Publications Sage CA: Los Angeles, CA}
+  publisher={SAGE Publications Sage CA: Los Angeles, CA},
+  doi={https://doi.org/10.1177/2332858416657343}
 }
 
 @article{bassok2016,
@@ -25,7 +26,8 @@
   number={3},
   pages={2332858416653924},
   year={2016},
-  publisher={SAGE Publications Sage CA: Los Angeles, CA}
+  publisher={SAGE Publications Sage CA: Los Angeles, CA},
+  doi={https://doi.org/10.1177/2332858416653924}
 }
 
 @article{chmielewski2016,
@@ -36,7 +38,8 @@
   number={3},
   pages={2332858416649593},
   year={2016},
-  publisher={SAGE Publications Sage CA: Los Angeles, CA}
+  publisher={SAGE Publications Sage CA: Los Angeles, CA},
+  doi={https://doi.org/10.1177/2332858416649593}
 }
 
 @book{agresti2011,
@@ -62,7 +65,8 @@
   number={4},
   pages={415--441},
   year={1979},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={https://doi.org/10.2307/589632}
 }
 
 @Manual{rsoftware,
