### 2019-11-08 448ce5e
Adds minimum tidyr version
