### 2019-11-08 60208d6
Adds more informative message for ordered factor
