### 2019-11-08 5c0fe0d
Fixes image path from README
