### 2019-11-08 b78d2b0
Fixes typos + hyperlinks
