+++ man/perc_diff.Rd
@@ -30,7 +30,7 @@ percentiles to subtract}
 }
 \value{
 \code{perc_diff} returns a vector with the percentile difference and
-it's associated standard error. \code{perc_diff_df} returns the same but as
+its associated standard error. \code{perc_diff_df} returns the same but as
 a data frame.
 }
 \description{
