### 2019-11-08 cimentadaj/perccalc#2 @cimentadaj
Thank you both for all of the comments, they've made the package much better. To make it easier for you, I've created a new branch `joss_review` [here](https://github.com/cimentadaj/perccalc/tree/joss_review) where you can find all of the commits that address your comments. I will address each of your comments and refer to the commit that implements it whenever possible.

In response to @amoeba's comments

- [X] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
Commit: 9cb93a653730ab9b458fe0eaca4235b706204c34

Introduced the correct MIT license file

Commit: 7b547a4cc37f5c70f4be7a6acd3e802e1e05024a

As suggested by @amoeba, I borrow some of the explanations from the paper as they are much better clearer there. This is slightly lengthier but explains in detail what the package does and what type of problem it fixes.

Commit: 0c1b11d2535e3255a35118789203525e1775ded9

I've rewritten the entire example using the GSS data and tackling a real world problem. In the example I try to explain exactly how the package addresses the problem.

- [X] The paper includes detail in the Examples section that JOSS guidelines indicate should be removed, see Guidelines: "Note the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.". Please remove this section.
Commit: 9f8e8298653244fead5303c27743651fd1eeb6c3

I'm not completely sure this change fixes **all** API discussion but I removed all references to what each function returns or details about the functionality of the functions in the example section. I'm happy to review this further if you think I'm missing some specifics.

- [X] The footer has a malformed citation
Commit: 4d09f209f6ec5b00c6e2907a309b69aaaa54a013

So useful of you to look up the DOI and thank you very much for pointing this out.

- [X] I got a warning on this section of the example in the README:
```r
        perc_diff(df, categorical, continuous, percentiles = c(90, 10))
        difference se
        79.61141798 0.08179896
        cols is now required.
        Please use cols = c(value)
```
Commit: 40ef5691b0cdaad63d6d52736d82cfc5827789f4

You're getting the warning because you installed `perccalc` from CRAN, which still has the `tidyr` dependency. In the Github version I've removed almost all dependencies and this is now fixed. I haven't submitted the package to CRAN because I was waiting for this review to finish, so this message should not occur with `devtools::install_github("cimentadaj/perccalc")`. However, in the commit above I did add a minimum version requirement for `tidyr` in the `Suggests` list of the `DESCRIPTION`.

- [X] This errors out in an odd way and could be a more descriptive errors if you think it'd help users:

```r
    perc_diff(df, foo, bar, percentiles = c(90, 10))
    Error in category_summary(data_model, variable_name, continuous_name, :
    is_ordered_fct is not TRUE
```
Commit: 60208d686b63418436918ae587a345f82659befd

I added a more informative message in the commit.

- [X] Typo in "A vector with the percentile difference and it's associated standard error" in perc_diff. "it's" should be "its".
- [X] In the README, "Reardon (2011)" could be hyperlinked. It's down at the bottom of the readme but I like links when I'm on websites.
- [X] In Usage section of README, the first code block doesn't list library(perccalc). I get that this is listed above in the Installation section but I think it might be more clear to repeat it in this code block

The first two changes have been applied in b78d2b0167bedc72ede095824678235cb049ff95 and the second one in 7b547a4cc37f5c70f4be7a6acd3e802e1e05024a

As outlined in the comments, if there's something that you still think that needs more explanation from my part, I'm happy to tackle this. Thanks for taking the time to review the package.
