### 2019-11-08 1cf7bca
Fixes typos
