### 2019-11-08 7fe5252
Adds paragraph on other pkgs using categorical methods
