+++ paper/paper.bib
@@ -102,3 +102,38 @@
   isbn = {978-3-319-24277-4},
   url = {https://ggplot2.tidyverse.org},
 }
+
+@Manual{boonstr,
+  title = {hbsae: Hierarchical Bayesian Small Area Estimation},
+  author = {Harm Jan Boonstra},
+  year = {2012},
+  note = {R package version 1.0},
+  url = {https://CRAN.R-project.org/package=hbsae},
+}
+
+@Article{vanbuuren,
+  title = {{mice}: Multivariate Imputation by Chained Equations in R},
+  author = {Stef {van Buuren} and Karin Groothuis-Oudshoorn},
+  journal = {Journal of Statistical Software},
+  year = {2011},
+  volume = {45},
+  number = {3},
+  pages = {1-67},
+  url = {https://www.jstatsoft.org/v45/i03/},
+}
+
+@Misc{ordinal,
+  title = {ordinal---Regression Models for Ordinal Data },
+  author = {R. H. B. Christensen},
+  year = {2019},
+  note = {R package version 2019.4-25. http://www.cran.r-project.org/package=ordinal/},
+}
+
+@Manual{gombin,
+  title = {qualvar: Implements Indices of Qualitative Variation Proposed by Wilcox
+(1973)},
+  author = {Joel Gombin},
+  year = {2018},
+  note = {R package version 0.2.0},
+  url = {https://CRAN.R-project.org/package=qualvar},
+}
