### 2019-11-08 cimentadaj/perccalc#2 @cimentadaj
In response to @briatte's comments:

- [X] I have left this item unchecked after reading @amoeba's comment, which points to the fact that the current LICENSE is nonstandard. I've submitted a PR: #3
Commit: 9cb93a653730ab9b458fe0eaca4235b706204c34

I fixed this in 9cb93a653730ab9b458fe0eaca4235b706204c34 before I knew that you had issued a PR. Otherwise I would've accepted yours :+1:.

- [X] Suggestion -- would it be better to use remotes instead of devtools (which uses remotes internally) in the installation code?

Hmm, I think it really doesn't matter for the package itself. None of the `devtools` calls in the package are **executed** (they're all in `Rmd` files and `eval` is set to `FALSE`) so `devtools` is not a dependency of `perccalc`. It would make sense to use `remotes` if I would be adding `devtools` as a dependency but fortunately, that can be avoided by not evaluating any call. Thanks for the suggestion though, I hadn't even though about this dependency problem (I even thought `devtools` **would be** a dependency even if you didn't evaluate any chunks of it).

- [X] Also -- like @amoeba, I favour having installation code in the "Install" section of the README, and then have the "Usage" section start with a library call.
Commit: 7b547a4cc37f5c70f4be7a6acd3e802e1e05024a

Implemented in the commit above.

- [X] Question -- is it now standard to have the CONTRIBUTING guide in .github instead of in the root of the repo?

I'm have no idea to be honest. I can't even remember how/when I added this CONTRIBUTING file. I don't recall doing this manually so it's probably taken from somewhere like `usethis` or something like that.

- [X] I've submitted some PRs to it:
#4 (suggested change in first sentence)
#5 (typos)
#6 (code suggestion)

Thank you very much for the PRs! I wouldn't have noticed these small things. I've accepted all of them in the `master` branch, although all changes related to this are being done in the branch `joss_review`. I'll make sure there aren't any conflicts once I merge `joss_review` into `master`.

- [X] The code in the paper won't run on old versions of tidyr -- specifically, line 145 (copied below) will fail on e.g. tidyr 0.8.3, which is what I had installed before updating it to 1.0.0.
Commit: 40ef5691b0cdaad63d6d52736d82cfc5827789f4

I've added a minimum `tidyr` version to the `Suggests` field in `DESCRIPTION`. This should only have some effect on the vignettes or README since `tidyr` is now removed as a dependency from `perccalc`.

- [X] (values of a continuous variable for the 0, 10th, 20th, ..., 100th percentile of the ordered categorical variable): … That should be either 0, 10th, …, 99th or 1st, 10th, …, 100th -- right?
- [X] Mathematics: … Do you need the capital M in English?

Commit: 1cf7bca338b1df4294f3f7d6b9f326abc8dbf661

Thank you for spotting them. Both typos are fixed in the commit above.

- [X] I have not yet checked that box because, unless I missed a paragraph somewhere in the paper, the package does not refer to other packages dedicated to working with categorical variables, e.g. qualvar (disclosure: the author is a friend of mine).
Commit: 7fe52520901069d6c4d75aad478d52db998ca8d4

Thanks for bringing this up -- I wasn't aware of `qualvar` and it is very much in the same area as `perccalc`. I've included a paragraph with a general overview of categorical-based methods in the paper.

- [X] The package has a startup message: Please cite as: Cimentada, Jorge (2019). Estimate Percentiles from an Ordered Categorical Variable R package version 1.0.3.

Yeah, this is a bit a tricky to refute. This is just an attempt to remind users that the package has indeed a paper attached to it (when published in `JOSS` this will change) and to prime them to recognize this work as academic work as well. This is clearly a contentious topic and I'm experimenting with it in two of my packages. I would prefer to leave it as is for the mean time.
