### 2019-11-08 #1796 @cimentadaj
I've answered to all comments from both reviewers [here](https://github.com/cimentadaj/perccalc/issues/2). As I outlined there, I'm happy to review some points if they feel like their points weren't addressed.

Thank you both for reviewing the paper/package, the comments have been very helpful to ship the next version to CRAN flawlessly!
