### 2019-11-08 cimentadaj/perccalc#2 @briatte
> @briatte sorry just saw your comment re DOI links- I can push your suggestion up the chain; thanks!

Please do :)
