### 2019-11-08 cimentadaj/perccalc#2 @briatte
Re: @cimentadaj's messages -- the replies above make sense to me: I'd like to approve the package at that stage.
