### 2019-11-09 cimentadaj/perccalc#2 @amoeba
Thanks for your detailed response to my review @cimentadaj ! It made it easy to inspect your work and verify the changes (which are all great, btw).

Regarding the edits you made the paper, I think it would be appropriate to remove the entire "Example" section in order to better match JOSS' standard. Please take a look at https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper and see if you think my interpretation of those guidelines is correct. Also, look at the most recently published paper for an example: https://joss.theoj.org/papers/10.21105/joss.01642. What do you think?
