### 2019-11-13 #1796 @majensen
@briatte, @amoeba -- based on @cimentadaj comments at [issue 2](https://github.com/cimentadaj/perccalc/issues/2), are you guys prepared to check off the remaining boxes?
Assuming so, @cimentadaj, would you like to merge your review branch into master?
