### 2019-11-13 cimentadaj/perccalc#2 @cimentadaj
Thanks for the suggestion @amoeba. I wasn't fully aware that examples were not allowed; in fact, I thought they were encouraged.  Taken from the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#example-usage): The authors should include examples of how to use the software (ideally to solve real-world analysis problems).

But I see where you confusion arises from. I'm not entirely sure whether they mean **written** examples or examples that actually use **code**. However, I'm inclined to think that they're ambiguous on this specific point (as in they allow both **written** or **code** examples) because recent papers also include code examples ([here](https://joss.theoj.org/papers/10.21105/joss.01798), [here](https://joss.theoj.org/papers/10.21105/joss.01857), [here](https://joss.theoj.org/papers/10.21105/joss.01722)).

If they allow both types of examples, then I'd prefer keeping the example section as I think it gives a very specific idea of how to solve a real-world problem with the package. This is not meant to be an exhaustive tutorial; for that, the user can read the vignettes of the package which are slightly more elaborate.

Perhaps @briatte has an opinion on this?
