### 2019-11-13 cimentadaj/perccalc#2 @briatte
> Perhaps @briatte has an opinion on this?

Not sure about what JOSS is really saying either, but personally, I think if there's a possibility of having a worked-out example, "(ideally to solve real-world analysis problems)", as per the review criteria, then we should use it.
