### 2019-11-13 cimentadaj/perccalc#2 @cimentadaj
@amoeba, if you think this might be important then we should get the editor involved to see if he can clarify what we should do.
