### 2019-11-14 cimentadaj/perccalc#2 @amoeba
Hey @cimentadaj, @briatte. I'm okay with whatever JOSS says on this. Though, the phrase "The authors should include examples of how to use the software (ideally to solve real-world analysis problems)" is from the Documentation section, not the Paper section of the guidelines, hence my interpretation.

Feel free to bump this up the chain @cimentadaj.
