### 2019-11-14 cimentadaj/perccalc#2 @cimentadaj
@majensen, we're very close to finishing the review but we're confused as to whether JOSS would be okay with the current example I provide in the paper. The example contains a brief code example of what the package works but we're unsure whether JOSS encourages such examples. Can you clarify this for us? If it's ok, then we're done with the review.
