### 2019-11-19 cimentadaj/perccalc#2 @majensen
Thanks @cimentadaj -- @arfon, @labarba would you please chime in here? I personally have no problem with either location, but my experience is still limited as yet.
