### 2019-11-22 cimentadaj/perccalc#2 @arfon
> Thanks @cimentadaj -- @arfon, @labarba would you please chime in here? I personally have no problem with either location, but my experience is still limited as yet.

As the example is rather long (it seems to take the paper to ~6 pages) I'd recommend removing this and perhaps linking to the examples where the are in the docs?
