### 2019-11-23 cimentadaj/perccalc#2 @briatte
> @cimentadaj - per @arfon 's [comment](https://github.com/cimentadaj/perccalc/issues/2#issuecomment-557629955) and @amoeba 's assessment, can we move the full example to docs and make a brief reference to it in the paper (maybe w/ a linkout)? I would think at that point we're good to go.

This looks like a good solution to me. I also think we're good to go on this package -- the review process did not find anything that critically needs fixing.
