### 2019-11-25 9919b23
Removes example section
