+++ paper/paper.Rmd
@@ -60,7 +60,7 @@ technique is not limited to this field alone and can be used essentially in any
 context where percentiles of ordered categorical variables are of interest. One
 example where this would provide useful would be in medicine based research,
 where demographic characteristics such as education categories are common
-factors for looking at differences between groups. 
+factors for looking at differences between groups.
 
 The field of computational categorical data analysis has a long history in R
 with packages addressing small-area estimation for categorical variables
@@ -78,159 +78,12 @@ Currently, `perccalc` implements:
 * Calculating differences in a continuous variable relative to the percentiles
   of an ordered categorical variable
 * Calculating values for a continuous variable relative to the percentiles of an
-  ordered categorical variable (values of a continuous variable for the 1th, 10th,
-  20th, ..., 100th percentile of the ordered categorical variable)
+  ordered categorical variable (values of a continuous variable for the 1th,
+  10th, 20th, ..., 100th percentile of the ordered categorical variable)
 * Weight-adjusted estimations for all percentile calculations
 * Provides uncertainty estimates for all calculations which allows the user to
   produce uncertainty intervals or propagate further calculations with these
   uncertainty coefficients
-
-Below I introduce the reader to one simplified example of `perccalc`'s
-capabilities in a real world scenario.
-
-# Example
-
-For this example, we will calculate the percentile difference in Mathematics
-test scores based on the education of the parents for several countries using
-the PISA 2006 and PISA 2012 datasets. Let's load `perccalc` with the packages
-`dplyr` [@dplyr], `tidyr` [@tidyr] and `ggplot2` [@ggplot2] for wrangling and
-visualizing the data:
-
-```{r, message = FALSE}
-library(perccalc)
-library(tidyr)
-library(ggplot2)
-library(dplyr)
-```
-
-`percalc` automatically loads `pisa_2012` and `pisa_2006` which are two datasets
-with all the information that we need. These two datasets have data for Estonia,
-Germany and Spain and contain the average test score in Mathematics for each
-student and their father's education as measured by the international ISCED
-classification. The first thing we have to do is make sure the categories are
-ordered.
-
-```{r setup}
-order_edu <- c("None", "ISCED 1", "ISCED 2",
-               "ISCED 3A, ISCED 4", "ISCED 3B, C",
-               "ISCED 5A, 6", "ISCED 5B")
-
-# Make ordered categories of our categorical variables
-pisa_2006 <-
-  pisa_2006 %>%
-  mutate(father_edu = factor(father_edu, levels = order_edu, ordered = TRUE))
-
-pisa_2012 <-
-  pisa_2012 %>%
-  mutate(father_edu = factor(father_edu, levels = order_edu, ordered = TRUE))
-
-# Merge them together
-pisa <- rbind(pisa_2006, pisa_2012)
-```
-
-Once the categories are ordered, `perc_diff` can calculate the percentile
-difference between the 90th and 10th percentile for the complete sample. For
-example:
-
-```{r }
-perc_diff(data_model = pisa,
-          categorical_var = father_edu,
-          continuous_var = avg_math,
-          percentiles = c(90, 10)
-          )
-```
-
-This means that the difference in Mathematics test scores between the 90th and
-10th percentile of father's education is 45 points with a standard error of 14
-points. We can extrapolate this example for each country separately using `perc_diff_df`
-which returns a data frame:
-
-```{r }
-cnt_diff <-
-  pisa %>%
-  nest(data = c(-year, -CNT)) %>%
-  mutate(
-    # Look through each year/cnt combination and apply perc_diff_df
-    edu_diff = lapply(data, function(x) {
[…]
