### 2019-11-25 1ddd95b
Removes horizontal break
