### 2019-11-25 cimentadaj/perccalc#2 @cimentadaj
I've removed the example section from the paper and indicated that there are a full-featured examples in the package's website. I think we should be good to go after this. @majensen I think we're done here!
