### 2019-11-25 cimentadaj/perccalc#2 @amoeba
@cimentadaj looks good to me. I have no outstanding, unresolved review items left so my review is now an _Accept_. I'll go indicate as much over on https://github.com/openjournals/joss-reviews/issues/1796 too.
