### 2019-11-25 #1796 @amoeba
@majensen my review is now an _Accept_ after @cimentadaj 's recent change to the paper, removing the examples, as per https://github.com/cimentadaj/perccalc/issues/2#issuecomment-558031545.
