### 2019-11-28 #1796 @majensen
@amoeba and @briatte thanks very much; and thanks @cimentadaj for your hard work and responsiveness.
@briatte, can you review your checklist and make sure everything is checked off; there are a couple of ticks necessary and I don't want to assume.

I will perform my final proofreading tasks - might lead to a PR - and then will make the formal recommendation.
