### 2019-11-28 cimentadaj/perccalc#2 @majensen
@cimentadaj would you like to merge joss_review into master?
