+++ paper/paper.Rmd
@@ -85,165 +85,9 @@ Currently, `perccalc` implements:
   produce uncertainty intervals or propagate further calculations with these
   uncertainty coefficients
 
-<<<<<<< HEAD
-`perccalc` offers flexibility and reliability for estimating any
-=======
-Below I introduce the reader to one simplified example of `perccalc`'s
-capabilities in a real world scenario.
-
-# Example
-
-For this example, we will calculate the percentile difference in mathematics
-test scores based on the education of the parents for several countries using
-the PISA 2006 and PISA 2012 datasets. Let's load `perccalc` with the packages
-`dplyr` [@dplyr], `tidyr` [@tidyr] and `ggplot2` [@ggplot2] for wrangling and
-visualizing the data:
-
-```{r, message = FALSE}
-library(perccalc)
-library(tidyr)
-library(ggplot2)
-library(dplyr)
-```
-
-`percalc` automatically loads `pisa_2012` and `pisa_2006` which are two datasets
-with all the information that we need. These two datasets have data for Estonia,
-Germany and Spain and contain the average test score in Mathematics for each
-student and their father's education as measured by the international ISCED
-classification. The first thing we have to do is make sure the categories are
-ordered.
-
-```{r setup}
-order_edu <- c("None", "ISCED 1", "ISCED 2",
-               "ISCED 3A, ISCED 4", "ISCED 3B, C",
-               "ISCED 5A, 6", "ISCED 5B")
-
-# Make ordered categories of our categorical variables
-pisa_2006 <-
-  pisa_2006 %>%
-  mutate(father_edu = factor(father_edu, levels = order_edu, ordered = TRUE))
-
-pisa_2012 <-
-  pisa_2012 %>%
-  mutate(father_edu = factor(father_edu, levels = order_edu, ordered = TRUE))
-
-# Merge them together
-pisa <- bind_rows(pisa_2006, pisa_2012)
-```
-
-Once the categories are ordered, `perc_diff` can calculate the percentile
-difference between the 90th and 10th percentile for the complete sample. For
-example:
-
-```{r }
-perc_diff(data_model = pisa,
-          categorical_var = father_edu,
-          continuous_var = avg_math,
-          percentiles = c(90, 10)
-          )
-```
-
-This means that the difference in Mathematics test scores between the 90th and
-10th percentile of father's education is 45 points with a standard error of 14
-points. We can extrapolate this example for each country separately using
-`dplyr::group_by`. Since the result of `perc_diff` is a vector, to be able to
-work seamlessly with `dplyr::group_by` we will use `perc_diff_df` which returns
-a data frame:
-
-```{r }
-cnt_diff <-
-  pisa %>%
-  nest(data = c(-year, -CNT)) %>%
-  mutate(
-    # Look through each year/cnt combination and apply perc_diff_df
-    edu_diff = lapply(data, function(x) {
-      perc_diff_df(data_model = x,
-                   categorical_var = father_edu,
-                   continuous_var = avg_math,
-                   percentiles = c(90, 10))
-    })
-  ) %>%
-  select(-data) %>% 
-  unnest(edu_diff)
-
-title_y_axis <- "90th/10th percentiles gap in Mathematics"
-
-# Plot results
-cnt_diff %>% 
-  ggplot(aes(year, difference, color = CNT, shape = CNT)) +
-  geom_point() +
-  geom_line() +
-  theme_minimal() +
-  scale_y_continuous(name = title_y_axis) +
-  scale_x_continuous(name = "Year")
-```
-
-It looks like Estonia has a much smaller achievement gap relative to Spain and
-Germany but also note that both Germany and Spain have been decreasing their
-inequality.
-
[…]
