### 2019-11-28 cimentadaj/perccalc#2 @cimentadaj
Should be merged now.

-----------------------------------

Jorge Cimentada
*https://cimentadaj.github.io/ <https://cimentadaj.github.io/>*

On Thu, Nov 28, 2019 at 4:01 AM Mark Jensen <notifications@github.com>
wrote:

> @cimentadaj <https://github.com/cimentadaj> would you like to merge
> joss_review into master?
>
