### 2019-11-28 7616f09
ed: Minor edits for grammar and readability
