### 2019-11-28 #1796 @majensen
@cimentadaj : created https://github.com/cimentadaj/perccalc/pull/7 with some minor edits
