### 2019-11-29 #1796 @majensen
@cimentadaj - Here are the final steps. Can I get you to create a tag for the latest master, and then archive that tag with [Zenodo](https://zenodo.org/) or similar? Then please provide the DOI from the archive in this thread. I will push the recommendation from there to the editors-in-chief.
Thanks!!
