### 2019-12-01 #1796 @majensen
(BTW, here's an overview for Zenodo, if you haven't used it before: https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729)
