### 2019-12-02 #1796 @cimentadaj
Thanks, the DOI is 10.5281/zenodo.3559855 and can be found [here](https://zenodo.org/record/3559855).
