### 2019-12-02 #1796 @majensen
@openjournals/joss-eics, I recommend this paper for publication and ask that you begin the final steps. Thanks!
