### 2019-12-02 #1796 @danielskatz
@majensen - if you can tell whedon the archive and version, that helps the AEiC...  (See https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance)
