### 2019-12-02 #1796 @danielskatz
👋 @cimentadaj - please add the DOI (https://doi.org/10.18637/jss.v045.i03) for the van Buuren paper to you bib file.  Also, should there be some link for the reference to Ordinal?
