### 2019-12-03 d973051
Fixes agresti DOI citation problem
