+++ paper/paper.bib
@@ -41,12 +41,14 @@
   publisher={SAGE Publications Sage CA: Los Angeles, CA},
   doi={https://doi.org/10.1177/2332858416649593}
 }
-
-@book{agresti2011,
-  title={Categorical data analysis},
-  author={Agresti, Alan and Kateri, Maria},
-  year={2011},
-  publisher={Springer}
+                  
+@misc{agresti2007,
+  title={An Introduction to Categorical Data Analysis},
+  url={http://dx.doi.org/10.1002/0470114754},
+  DOI={10.1002/0470114754},
+  publisher={John Wiley & Sons, Inc.},
+  author={Agresti, Alan},
+  year={2007}
 }
 
 @book{agresti2010,
@@ -120,6 +122,7 @@
   number = {3},
   pages = {1-67},
   url = {https://www.jstatsoft.org/v45/i03/},
+  doi = {https://doi.org/10.18637/jss.v045.i03}  
 }
 
 @Misc{ordinal,
