### 2019-12-03 #1796 @cimentadaj
Thanks @danielskatz, should be fixed now for the van Buuren paper and the Agresti book.
