### 2019-12-03 #1796 @danielskatz
There are a number of other small changes, which I've suggested fixes for in https://github.com/cimentadaj/perccalc/pull/8 - please merge, or let me know which you disagree with.
