### 2019-12-03 #1796 @danielskatz
👋 @briatte - There are a number of checkboxes that you didn't check.  Can you check them please, or explain why they can't yet be checked?
