### 2019-12-06 #1796 @majensen
@danielskatz : My guess is @briatte is OOO, since he is usu. very responsive. In view of comment from @briatte 's comment at https://github.com/openjournals/joss-reviews/issues/1796#issuecomment-558501728 and previous, I would recommend moving forward with the publication.
